+++ packages/ubq/CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
-
+ - `application.customSettings.mailerConfig` section now defined in model partial config
+ and automatically merged into main config (starting from ub@5.18.12). See README.md for 
+ environment variables list.
+ 
 ### Deprecated
 
 ### Removed